Reorganize project
This is a significant overhaul of the project. The most important change
is that testing is isolated from the main tool. The tests also expect
valid inputs to test against. Testers must specify a valid CIS PDF and
an assertion file for comparison, which get fed into the test class.